# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexoU API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/reports": {
      "post": {
        "summary": "Crear reporte pendiente",
        "description": "Solo estudiantes. El servidor determina autor, UUID, folio y fechas. También admite JSON sin foto.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateReport"
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "titulo": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "descripcion": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "areaId": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "categoriaId": {
                    "type": "integer",
                    "minimum": 1
                  }
                },
                "required": [
                  "titulo",
                  "descripcion",
                  "areaId",
                  "categoriaId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Correcto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o inaccesible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Carga excesiva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Reportes autorizados paginados",
        "description": "Orden descendente (creado_en,id). Estudiante solo propios; personal listado general.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "description": ""
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500
            },
            "description": "Cursor de la página anterior; conservar filtros."
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pendiente",
                "revision",
                "solucionado"
              ]
            },
            "description": ""
          },
          {
            "in": "query",
            "name": "own",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Forzar propios incluso para personal."
          },
          {
            "in": "query",
            "name": "areaId",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": ""
          },
          {
            "in": "query",
            "name": "categoriaId",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": ""
          },
          {
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": ""
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Correcto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPage"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o inaccesible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Carga excesiva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reports/summary": {
      "get": {
        "summary": "Conteos completos por estado",
        "description": "",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "own",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Correcto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o inaccesible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Carga excesiva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{id}": {
      "get": {
        "summary": "Detalle autorizado e historial",
        "description": "statusUpdates excluye el evento inicial de creación.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Correcto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o inaccesible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Carga excesiva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{id}/status": {
      "patch": {
        "summary": "Cambiar estado con nota",
        "description": "Solo personal. Rutina transaccional, reapertura permitida; estado repetido no añade evento.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Correcto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o inaccesible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Carga excesiva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{id}/evidence": {
      "get": {
        "summary": "Evidencia privada",
        "description": "",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Foto autorizada; Cache-Control: private, no-store",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente/expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o inaccesible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Carga excesiva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "error",
          "requestId"
        ]
      },
...
      "StatusUpdate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "ID opaco; los BIGINT se representan como cadenas."
          },
          "from": {
            "type": "string",
            "enum": [
              "pendiente",
              "revision",
              "solucionado"
            ]
          },
          "to": {
            "type": "string",
            "enum": [
              "pendiente",
              "revision",
              "solucionado"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string",
            "description": "ID opaco; los BIGINT se representan como cadenas."
          },
          "authorName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "from",
          "to",
          "note",
          "createdAt",
          "authorId",
          "authorName"
        ]
      },
      "Report": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "ID opaco; los BIGINT se representan como cadenas."
          },
          "folio": {
            "type": "string"
          },
          "ownerId": {
            "type": "string",
            "description": "ID opaco; los BIGINT se representan como cadenas."
          },
          "ownerNombre": {
            "type": "string"
          },
          "titulo": {
            "type": "string",
            "maxLength": 80
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500
          },
          "areaId": {
            "type": "string",
            "description": "ID opaco; los BIGINT se representan como cadenas."
          },
          "area": {
            "type": "string"
          },
          "categoriaId": {
            "type": "string",
            "description": "ID opaco; los BIGINT se representan como cadenas."
          },
          "categoria": {
            "type": "string"
          },
          "estado": {
            "type": "string",
            "enum": [
              "pendiente",
              "revision",
              "solucionado"
            ]
          },
          "evidenceUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "statusUpdates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusUpdate"
            }
          }
        },
        "required": [
          "id",
          "folio",
          "ownerId",
          "ownerNombre",
          "titulo",
          "descripcion",
          "areaId",
          "area",
          "categoriaId",
          "categoria",
          "estado",
          "evidenceUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateReport": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "titulo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "descripcion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "areaId": {
            "type": "integer",
            "minimum": 1
          },
          "categoriaId": {
            "type": "integer",
            "minimum": 1
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "JPEG/PNG/WebP decodificable, hasta MAX_PHOTO_BYTES (por defecto 5 MiB). MIME comprobado por contenido."
          }
        },
        "required": [
          "titulo",
          "descripcion",
          "areaId",
          "categoriaId"
        ]
      },
      "ChangeStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "pendiente",
              "revision",
              "solucionado"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          }
        },
        "required": [
          "estado"
        ]
      },
      "Summary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "pendiente": {
            "type": "integer",
            "minimum": 0
          },
          "revision": {
            "type": "integer",
            "minimum": 0
          },
          "solucionado": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "total",
          "pendiente",
          "revision",
          "solucionado"
        ]
      },
...
      "ReportPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Report"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "nextCursor",
          "total"
        ]
      },
...
    }
  }
}